universefy service file strutcture update
