+++ services/universefy/services/index.js
@@ -1,12 +1,1 @@
-const generateRandomString = (length) => {
-	let text = '';
-	const possible = 'ABCDEFGHIJKLMNOPQRSTUVWXYZabcdefghijklmnopqrstuvwxyz0123456789';
-
-	for(let i = 0; i < length; i++) {
-		text += possible.charAt(Math.floor(Math.random() * possible.length));
-    }
-    
-	return text;
-};
-
-exports.generateRandomString = generateRandomString;
+exports.generateRandomString = require('./generateRandomString');
